feat(agent): stamp tool results with absolute start/end times
event.Tool gains StartedAt/EndedAt (unix ms), captured around the actual
execution in executeBatch and mirrored on the wire as startedAt/endedAt.
Calls that never ran (dependency-skipped, cancelled, synthetic) carry
zero so consumers can tell 'fast' from 'never executed'.

To stay inside the file-size ratchet, the batch-execution cluster
(executeBatch, batch barriers, partitioning, runParallel) moves from
agent.go to execute_batch.go, with its carried-over comments tightened
to the repolint limits.
